moving operation fyle to model package
