+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/operations/components/OperationsListCard.kt
@@ -64,9 +64,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
+import br.com.rbrthmn.model.OperationType
 import br.com.rbrthmn.ui.financialcompanion.common.ReservesDropdownMenu
 import br.com.rbrthmn.ui.financialcompanion.screens.operations.Operation
-import br.com.rbrthmn.ui.financialcompanion.utils.OperationType
 import br.com.rbrthmn.ui.financialcompanion.utils.getOperationsMock
 import java.text.SimpleDateFormat
 import java.util.Date
