+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/reserves/ReservesScreen.kt
@@ -67,7 +67,7 @@ import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
 import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.AddOperationDialog
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
-import br.com.rbrthmn.ui.financialcompanion.utils.OperationType
+import br.com.rbrthmn.model.OperationType
 import br.com.rbrthmn.ui.financialcompanion.utils.valueWithCurrencyString
 import java.text.SimpleDateFormat
 import java.util.Locale
